+++ th-frontend/th-ng-configuration/src/app/dashboard/user/user.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort, MatPaginator } from '@angular/material';
-import { Path, User } from 'th-ng-commons';
+import { Path, User, ERole } from 'th-ng-commons';
 import { UserService } from '../../services/user.service';
 import { ToastrService } from 'ngx-toastr'
 @Component({
@@ -13,7 +13,7 @@ export class UserComponent implements OnInit {
 
   ready: boolean = false;
   dataSource: MatTableDataSource<any>;
-  displayedColumns: string[] = ['position', 'name', 'email', 'actions'];
+  displayedColumns: string[] = ['position', 'name', 'email', 'role', 'actions'];
 
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
@@ -34,7 +34,7 @@ export class UserComponent implements OnInit {
     this.userService.getAll().subscribe((users: User[]) => {
       let i = 1;
       const data: any[] = [];
-
+      users = users && users.length > 0 ? users : [];
       for (const a of users) {
         const obj = Object.assign({ position: 0 }, a);
         obj.position = i;
@@ -56,18 +56,36 @@ export class UserComponent implements OnInit {
   }
 
   delete(id) {
-    this.userService.delete(id).subscribe(res => {
-      if (res['deleted']) {
-        this.toastr.success('Usuario eliminado correctamente');
-        this.getData();
-      } else {
-        this.toastr.warning('No es posible eliminar el usuario');
-      }
-    }, error => this.toastr.warning('No es posible eliminar el usuario'));
+    if (confirm('El elemento seleccionado será eliminado ¿deseas continuar?')) {
+      this.userService.delete(id).subscribe(res => {
+        if (res['deleted']) {
+          this.toastr.success('Usuario eliminado correctamente');
+          this.getData();
+        } else {
+          this.toastr.warning('No es posible eliminar el usuario');
+        }
+      }, error => this.toastr.warning('No es posible eliminar el usuario'));
+    }
   }
 
   public get lPath(): Path[] {
     return [{ isActive: true, label: 'Usuarios', url: '' }];
   }
 
+  getRole(idrol: number): string {
+    switch (idrol) {
+      case ERole.ADMIN.key:
+        return ERole.ADMIN.value;
+        break;
+      case ERole.STUDENT.key:
+        return ERole.STUDENT.value;
+        break;
+      case ERole.EVALUATOR.key:
+        return ERole.EVALUATOR.value;
+        break;
+      default:
+        return "";
+        break;
+    }
+  }
 }
